sconcho now tracks the current project name.
The name is simply the name of the current file
and is now displayed in the window title. Also
extended the save function to an save/save as combo.
If a name is defined for the current file save
will simply save and otherwise pop up a save as
widget.