Streamlined canvas export process.
Exporting the legend now properly exports the legend only without
any surrounding blank grid areas!
Presently, relying and setting the isVisible attribute of the QtSvgItems
is unreliable and when grabbing QtGraphicsItems from the canvas
we should exclude them from the iteration range. Not quite sure why
that is at this point, maybe there is some sharing going on. We'll
have to investigate since the current workaround feels kind of
clunky.