Introduced new sconcho binary spf format.
Deprecated the old XML based spf format for a new
binary format to reduce file size and increase the
speed of saving. For now the code still supports
loading of the previous XML format but writes are
exclusively in the new binary spf format.

Also fixed a few small bugs in other parts of
the code.