preferenceDialog now changes label and legend fonts.
Also, I think I finally fixed the issue with delete
and insert columns/rows getting confused and doing
multiples. Since the dialog is persistent the
connections were not set up at initial setup only
but each time the dialog was opened which probably
confused things quite a bit. Let's hope that this
was it.