Reworked the way sconcho identifies symbols.
Previously, a symbol was recognized via category + name.
While this allowed symbols with the same name to exist in different
categories it broke saved projects if categories were
reshuffled since sconcho couldn't identify symbols any more.
Thus, the behavior was changed and symbols are now recognized by name only.
As long as users don't need to add multiple symbols of the same
name (which is unlikely) this should be fine.

Also added new symbols and re-categorized many of them in
addition to specifying a display order for all of them.