Completele reworks pattern repeat items.
PatternRepeatItems are now implemented via QGraphicsPolygonItems.
This has the advantage that they are aware of their shape
and properly detect canvas events. The previous implementation
in terms of a group of line items did not do that and
detection was basd on the bounding box which doesn't work
properly for non-rectangular repeats.