Re-implemented row/column labels
The new implementation of row/column labels stores the
PatternLabelObjects in a dictionary. Thus they don't
need to be re-instantiated all the time and can just
be updated. This also allows custom editable row/column
labels which will do the the right thing (i.e. remember
their text) when deleting/adding of rows and columns.
The custom row/column label option can be enabled in the
preferences.