Reworked IO for API 3 and added row column labels to spf files.
The new API 3 version of the spf file format has a completly
reworked layout. Each type of object to be written/read (patterGridItem,
labelItems, etc.) is now written into a separate section in the spf file
that starts off with its name and length (in number of objects).
Thus, sections can be parsed in any order. As a convention, the
settings section is written last but this is not required. Using
this approach the new API 3 format should be much more flexible
in accomodating new sections.

Finally cleaned up the code for API1 and API2 and separated API
specific functionality into separate functions.